\name{maize}
\alias{maize}
\docType{data}
\title{
Maize data
}
\description{
This is a dataset consisting of maize yields, environmental and management variables measured across 4 groups. The groups pertain to different seasons (2010 and 2013) for
different locations in Ethiopia (Pawe and Adami Tullu).
}
\usage{data("maize")}
\format{
  The format is:
List of 4
 $ :'data.frame':	82 obs. of  26 variables:
  ..$ season_days           : num [1:82] 184 184 184 184 184 184 184 184 184 184 ...
  ..$ TEMP_mean             : num [1:82] 21.2 21.2 21.2 21.2 21.2 ...
  ..$ RAD_sum               : num [1:82] 3.43e+09 3.43e+09 3.43e+09 3.43e+09 3.43e+09 ...
  ..$ ET0_sum               : num [1:82] 67.1 67.1 67.1 67.1 67.1 ...
  ..$ RAIN_sum              : num [1:82] 121 121 118 118 118 ...
  ..$ TEMP_cv               : num [1:82] 1554 1554 1554 1554 1554 ...
  ..$ biol_eff_degr_day     : num [1:82] 3905 3905 3905 3905 3905 ...
  ..$ wet_days              : num [1:82] 56 56 57 57 57 57 57 57 57 56 ...
  ..$ tropical_nights       : num [1:82] 1 1 1 1 1 1 1 1 1 1 ...
  ..$ max_consec_dry_days   : num [1:82] 74 74 74 74 74 74 74 74 74 74 ...
  ..$ max_consec_summer_days: num [1:82] 83 83 83 83 83 83 83 83 83 83 ...
  ..$ arid_days             : num [1:82] 88 88 84 84 84 84 84 84 84 88 ...
  ..$ sand                  : num [1:82] 313 313 306 306 306 306 306 308 296 279 ...
  ..$ silt                  : num [1:82] 167 167 153 153 153 139 139 117 118 153 ...
  ..$ clay                  : num [1:82] 445 445 447 447 447 469 469 470 467 445 ...
  ..$ cec                   : num [1:82] 143 143 149 149 149 148 148 147 144 157 ...
  ..$ cfvo                  : num [1:82] 215 215 223 223 223 222 222 224 229 229 ...
  ..$ nitrogen              : num [1:82] 59 59 59 59 59 59 59 59 60 59 ...
  ..$ phh2o                 : num [1:82] 293 293 287 287 287 280 280 275 279 272 ...
  ..$ soc                   : num [1:82] 263 263 266 266 266 251 251 255 254 283 ...
  ..$ ocd                   : num [1:82] 185 185 185 185 185 189 189 173 168 175 ...
  ..$ sowing_doy            : num [1:82] 169 169 169 169 169 169 169 169 169 169 ...
  ..$ harvest_doy           : num [1:82] 352 352 352 352 352 352 352 352 352 352 ...
  ..$ Yield                 : num [1:82] 2.401 0.931 1.377 4.401 1.545 ...
  ..$ soil_conservation     : num [1:82] 0 1 0 0 0 1 0 0 0 1 ...
  ..$ variety_OPV           : num [1:82] 0 0 0 0 0 0 0 0 0 0 ...
 $ :'data.frame':	82 obs. of  26 variables:
  ..$ season_days           : num [1:82] 184 184 184 184 184 184 184 184 184 184 ...
  ..$ TEMP_mean             : num [1:82] 21.2 21.2 21.2 21.2 21.2 ...
  ..$ RAD_sum               : num [1:82] 3.43e+09 3.43e+09 3.43e+09 3.43e+09 3.43e+09 ...
  ..$ ET0_sum               : num [1:82] 67.1 67.1 67.1 67.1 67.1 ...
  ..$ RAIN_sum              : num [1:82] 121 121 118 118 118 ...
  ..$ TEMP_cv               : num [1:82] 1554 1554 1554 1554 1554 ...
  ..$ biol_eff_degr_day     : num [1:82] 3905 3905 3905 3905 3905 ...
  ..$ wet_days              : num [1:82] 56 56 57 57 57 57 57 57 57 57 ...
  ..$ tropical_nights       : num [1:82] 1 1 1 1 1 1 1 1 1 1 ...
  ..$ max_consec_dry_days   : num [1:82] 74 74 74 74 74 74 74 74 74 74 ...
  ..$ max_consec_summer_days: num [1:82] 83 83 83 83 83 83 83 83 83 83 ...
  ..$ arid_days             : num [1:82] 88 88 84 84 84 84 84 84 84 84 ...
  ..$ sand                  : num [1:82] 313 313 306 306 306 306 306 306 306 297 ...
  ..$ silt                  : num [1:82] 167 167 153 153 153 153 153 153 153 115 ...
  ..$ clay                  : num [1:82] 445 445 447 447 447 447 447 447 447 459 ...
  ..$ cec                   : num [1:82] 143 143 149 149 149 149 149 149 149 149 ...
  ..$ cfvo                  : num [1:82] 215 215 223 223 223 223 223 223 223 218 ...
  ..$ nitrogen              : num [1:82] 59 59 59 59 59 59 59 59 59 59 ...
  ..$ phh2o                 : num [1:82] 293 293 287 287 287 287 287 287 287 283 ...
  ..$ soc                   : num [1:82] 263 263 266 266 266 266 266 266 266 258 ...
  ..$ ocd                   : num [1:82] 185 185 185 185 185 185 185 185 185 163 ...
  ..$ sowing_doy            : num [1:82] 164 164 165 165 165 165 165 165 165 165 ...
  ..$ harvest_doy           : num [1:82] 352 352 352 352 352 352 352 352 352 352 ...
  ..$ Yield                 : num [1:82] 0.001 1.668 0.801 2.801 2.401 ...
  ..$ soil_conservation     : num [1:82] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ variety_OPV           : num [1:82] 0 0 0 0 0 0 1 1 1 0 ...
 $ :'data.frame':	129 obs. of  26 variables:
  ..$ season_days           : num [1:129] 148 148 148 148 148 148 148 148 148 148 ...
  ..$ TEMP_mean             : num [1:129] 17.9 17.9 18.6 18.6 18.6 ...
  ..$ RAD_sum               : num [1:129] 3.02e+09 3.02e+09 3.01e+09 3.01e+09 3.01e+09 ...
  ..$ ET0_sum               : num [1:129] 61.3 61.3 62.7 62.7 62.7 ...
  ..$ RAIN_sum              : num [1:129] 50.5 50.5 50.2 50.2 50.2 ...
  ..$ TEMP_cv               : num [1:129] 1604 1604 1520 1520 1520 ...
  ..$ biol_eff_degr_day     : num [1:129] 2644 2644 2760 2760 2760 ...
  ..$ wet_days              : num [1:129] 23 23 23 23 23 23 23 23 23 23 ...
  ..$ tropical_nights       : num [1:129] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ max_consec_dry_days   : num [1:129] 27 27 27 27 27 27 27 27 27 27 ...
  ..$ max_consec_summer_days: num [1:129] 115 115 82 82 82 82 82 82 82 82 ...
  ..$ arid_days             : num [1:129] 117 117 118 118 118 118 118 118 118 118 ...
  ..$ sand                  : num [1:129] 366 366 426 426 426 418 423 423 432 432 ...
  ..$ silt                  : num [1:129] 33 33 37 37 37 42 22 22 37 37 ...
  ..$ clay                  : num [1:129] 351 351 248 248 248 282 239 239 243 243 ...
  ..$ cec                   : num [1:129] 185 185 165 165 165 170 160 160 162 162 ...
  ..$ cfvo                  : num [1:129] 225 225 202 202 202 197 205 205 201 201 ...
  ..$ nitrogen              : num [1:129] 71 71 73 73 73 73 73 73 74 74 ...
  ..$ phh2o                 : num [1:129] 261 261 320 320 320 310 336 336 339 339 ...
  ..$ soc                   : num [1:129] 389 389 432 432 432 408 426 426 418 418 ...
  ..$ ocd                   : num [1:129] 192 192 178 178 178 188 196 196 188 188 ...
  ..$ sowing_doy            : num [1:129] 165 165 165 165 165 165 165 165 165 165 ...
  ..$ harvest_doy           : num [1:129] 223 223 223 223 223 223 223 223 223 223 ...
  ..$ Yield                 : num [1:129] 1.131 1.301 1.201 0.401 0.745 ...
  ..$ soil_conservation     : num [1:129] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ variety_OPV           : num [1:129] 0 0 0 0 0 0 0 0 0 0 ...
 $ :'data.frame':	132 obs. of  26 variables:
  ..$ season_days           : num [1:132] 148 148 148 148 148 148 148 148 148 148 ...
  ..$ TEMP_mean             : num [1:132] 18.6 17.9 17.9 17.9 17.9 ...
  ..$ RAD_sum               : num [1:132] 3.01e+09 3.02e+09 3.02e+09 3.02e+09 3.02e+09 ...
  ..$ ET0_sum               : num [1:132] 62.7 61.3 61.3 61.3 61.3 ...
  ..$ RAIN_sum              : num [1:132] 51.7 50.5 50.5 50.5 50.5 ...
  ..$ TEMP_cv               : num [1:132] 1520 1604 1604 1604 1604 ...
  ..$ biol_eff_degr_day     : num [1:132] 2760 2644 2644 2644 2644 ...
  ..$ wet_days              : num [1:132] 22 23 23 23 23 23 23 23 23 23 ...
  ..$ tropical_nights       : num [1:132] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ max_consec_dry_days   : num [1:132] 27 27 27 27 27 27 27 27 27 27 ...
  ..$ max_consec_summer_days: num [1:132] 82 115 115 115 115 82 82 82 82 82 ...
  ..$ arid_days             : num [1:132] 118 117 117 117 117 118 118 118 118 118 ...
  ..$ sand                  : num [1:132] 426 390 390 366 366 426 426 426 426 426 ...
  ..$ silt                  : num [1:132] 43 40 40 33 33 37 37 37 37 37 ...
  ..$ clay                  : num [1:132] 215 383 383 351 351 248 248 248 248 248 ...
  ..$ cec                   : num [1:132] 161 147 147 185 185 165 165 165 165 165 ...
  ..$ cfvo                  : num [1:132] 197 200 200 225 225 202 202 202 202 202 ...
  ..$ nitrogen              : num [1:132] 72 72 72 71 71 73 73 73 73 73 ...
  ..$ phh2o                 : num [1:132] 355 276 276 261 261 320 320 320 320 320 ...
  ..$ soc                   : num [1:132] 430 341 341 389 389 432 432 432 432 432 ...
  ..$ ocd                   : num [1:132] 177 165 165 192 192 178 178 178 178 178 ...
  ..$ sowing_doy            : num [1:132] 165 165 165 165 165 165 165 165 165 165 ...
  ..$ harvest_doy           : num [1:132] 223 223 223 223 223 223 223 223 223 223 ...
  ..$ Yield                 : num [1:132] 0.43 1 1.47 4 4.1 ...
  ..$ soil_conservation     : num [1:132] 1 0 0 0 0 0 0 0 0 0 ...
  ..$ variety_OPV           : num [1:132] 0 0 0 0 0 0 0 0 0 0 ...
}
\details{
Contains a subset of data used in the Hermes et al. (2022) paper, which is a subset
of data used in the Vasco Silva et al. (forthcoming) paper.
}
\source{
1. Hermes, S., van Heerwaarden, J., and Behrouzi, P. (2022). Copula graphical models for heterogeneous mixed data. arXiv preprint, arXiv:2210.13140. \cr
}
\references{
1. Hermes, S., van Heerwaarden, J., and Behrouzi, P. (2022). Copula graphical models for heterogeneous mixed data. arXiv preprint, arXiv:2210.13140. \cr
2. Vasco Silva, J., J. van Heerwaarden, R. Pytrik, A. G. Laborte, K. Tesfaye, and M. K. van Ittersum (forthcoming). Big data, small explanatory power? lessons learnt with random forest predictive modeling of crop yield in contrasting farming systems. \cr
}
\examples{
data(maize)
}
\keyword{datasets}
